Add list of offices in main nav.
It would be nice if we didn’t have the series of data as the left nav, but rather the list of offices. Then, per office, show each series as a graph.

This should let us compare across series a bit better.

There aren’t any pages wired up to these entries, as the data set isn’t ordered to support it.